+++ NEWS.md
@@ -1,5 +1,7 @@
 # tidypredict (development version)
 
+- Fixed bug where `base_score` wasn't extracted correctly xgboost for version 3 or higher. (#173)
+
 # tidypredict 1.0.0
 
 ## Breaking Changes